+++ db/migrate/20130924091342_legacy_meeting_journal_data.rb
@@ -1,3 +1,4 @@
+#-- encoding: UTF-8
 #-- copyright
 # OpenProject is a project management system.
 # Copyright (C) 2011-2013 the OpenProject Foundation (OPF)
@@ -11,9 +12,20 @@
 #
 # See doc/COPYRIGHT.md for more details.
 #++
+#
 
-module OpenProject
-  module Meeting
-    VERSION = "3.0.3"
+require_relative 'migration_utils/legacy_journal_migrator'
+
+class LegacyMeetingJournalData < ActiveRecord::Migration
+  def up
+    migrator.run
+  end
+
+  def down
+    migrator.remove_journals_derived_from_legacy_journals 'meeting_journals'
+  end
+
+  def migrator
+    @migrator ||= Migration::LegacyJournalMigrator.new "MeetingJournal", "meeting_journals"
   end
 end